add Pool (No Swimming), add TouchHandler, remove background XML
Added a TouchHandler that handles the inputs on the view, the handler can handle multiple inputs but is configured so that one can fetch only one single input. The pool class is a helper class for the TouchHandler and is used to produce TouchEvents.

Removed the possibility to have XML in the background. Reason: the SurfaceView did not redraw properly (did not redraw the transparent pixels), this caused flashing objects onscreen and the cure for this would have been to replace the current view structure. I whent with the easy solution of removing the possiblity of an XML background.

Testing: I made a dummy class and added a drawn background to the GameContent class.