add Village, add LevelSelectActivity, add layouts for Village and levelSelect
Added a Village class containing values for total water amount and best run/ nr of runs, and so on. The village class is connected to the FileIO class and saved the data in the mobiles sharedPreferences.

Added a VillageActivity displaying the data from the Village class.

Added a levelSelectActivity, the activity allows you to select difficulties before starting a level.