add GameSettings, add AfterGameActivity, add Start and End
added a gameSettings class to controll difficulty levels and to define game time.

added a After game Activity wich is the one player end up in when they finnish the game. (containing image and information about water).

added a game time to the game, so that the game has a start and an end.