fix progressBar, fix WaterBar, add icons to AfterGameInfo, fix blue_water
Fixed the progress and water bar. They now correctly displays the person on the left and the water on the right in a way that is adjustable in the code. (Note: WaterBar still do not change color)

added icons in the After game menu to better display the "Total Water Amount". also added a button to the village.

fixed the blue_water image for the water bar. too mutch transparancy on Top and Bottom.